requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# AI/ML Dependencies
openai>=1.0.0
transformers>=4.36.0
sentence-transformers>=2.2.2
torch>=2.0.0

# LangChain Stack
langchain>=0.1.17
langchain-community>=0.0.36
langchain-core>=0.1.52
langchain-openai>=0.1.6

# Vector Store & Embeddings
faiss-cpu>=1.7.4
huggingface-hub>=0.20.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
requests>=2.31.0